backend/requirements.txt
pywinctl
pynput
pyautogui
pyperclip
google_genai
arrow
uvicorn
